+++ apps/meteor/client/startup/index.ts
@@ -20,7 +20,6 @@ import './reloadRoomAfterLogin';
 import './roles';
 import './rootUrlChange';
 import './routes';
-import './setupWizard';
 import './slashCommands';
 import './startup';
 import './streamMessage';