+++ apps/meteor/client/views/room/composer/hooks/useComposerBoxPopupQueries.ts
@@ -22,28 +22,31 @@ export const useComposerBoxPopupQueries = <T extends { _id: string; sort?: numbe
 			slashCommands.commands[(filter as any)?.cmd].providesPreview) ||
 		shouldPopupPreview;
 
+	const queries = useQueries({
+		queries: [
+			{
+				placeholderData: keepPreviousData,
+				queryKey: ['message-popup', 'local', filter, popup],
+				queryFn: () => popup?.getItemsFromLocal && popup.getItemsFromLocal(filter),
+				enabled: enableQuery,
+			},
+			{
+				placeholderData: keepPreviousData,
+				queryKey: ['message-popup', 'server', filter, popup],
+				queryFn: () => popup?.getItemsFromServer && popup.getItemsFromServer(filter),
+				enabled: counter > 0,
+			},
+		],
+	}) as QueriesResults<T[]>;
+
+	useEffect(() => {
+		if (Array.isArray(queries[0].data) && queries[0].data.length < 5) {
+			setCounter(1);
+		}
+	}, [queries]);
+
 	return {
-		queries: useQueries({
-			queries: [
-				popup?.getItemsFromLocal && {
-					placeholderData: keepPreviousData,
-					queryKey: ['message-popup', 'local', filter, popup],
-					queryFn: () => popup?.getItemsFromLocal && popup.getItemsFromLocal(filter),
-					onSuccess: (args: T[]) => {
-						if (args.length < 5) {
-							setCounter(1);
-						}
-					},
-					enabled: enableQuery,
-				},
-				popup?.getItemsFromServer && {
-					placeholderData: keepPreviousData,
-					queryKey: ['message-popup', 'server', filter, popup],
-					queryFn: () => popup?.getItemsFromServer && popup.getItemsFromServer(filter),
-					enabled: counter > 0,
-				},
-			].filter(Boolean) as any,
-		}) as QueriesResults<T[]>,
+		queries,
 		suspended: !enableQuery,
 	};
 };