+++ apps/meteor/ee/server/apps/marketplace/marketplaceErrors.ts
@@ -1,0 +1,13 @@
+export class MarketplaceAppsError extends Error {
+	constructor(message: string) {
+		super(message);
+		this.name = 'MarketplaceAppsError';
+	}
+}
+
+export class MarketplaceConnectionError extends Error {
+	constructor(message: string) {
+		super(message);
+		this.name = 'MarketplaceConnectionError';
+	}
+}