+++ apps/meteor/tests/end-to-end/api/livechat/09-visitors.ts
@@ -2,7 +2,6 @@ import { faker } from '@faker-js/faker';
 import type { ILivechatVisitor } from '@rocket.chat/core-typings';
 import { expect } from 'chai';
 import { before, describe, it, after } from 'mocha';
-import moment from 'moment';
 import { type Response } from 'supertest';
 
 import { getCredentials, api, request, credentials } from '../../../data/api-data';
@@ -482,27 +481,6 @@ describe('LIVECHAT - visitors', () => {
 					expect(res.body.visitor).to.have.property('ts');
 					expect(res.body.visitor._id).to.be.equal(createdVisitor._id);
 				});
-		});
-
-		it('should return visitor activity field when visitor was active on month', async () => {
-			// Activity is determined by a conversation in which an agent has engaged (sent a message)
-			// For a visitor to be considered active, they must have had a conversation in the last 30 days
-			const period = moment().format('YYYY-MM');
-			const { visitor, room } = await startANewLivechatRoomAndTakeIt();
-			// agent should send a message on the room
-			await request
-				.post(api('chat.sendMessage'))
-				.set(credentials)
-				.send({
-					message: {
-						rid: room._id,
-						msg: 'test',
-					},
-				});
-
-			const activeVisitor = await getLivechatVisitorByToken(visitor.token);
-			expect(activeVisitor).to.have.property('activity');
-			expect(activeVisitor.activity).to.include(period);
 		});
 
 		it('should not affect MAC count when a visitor is removed via GDPR', async () => {