+++ apps/meteor/tests/unit/server/livechat/hooks/markRoomResponded.spec.ts
@@ -4,7 +4,7 @@ import proxyquire from 'proxyquire';
 import Sinon from 'sinon';
 
 const models = {
-	LivechatVisitors: { isVisitorActiveOnPeriod: Sinon.stub(), markVisitorActiveForPeriod: Sinon.stub() },
+	LivechatContacts: { isContactActiveOnPeriod: Sinon.stub(), markContactActiveForPeriod: Sinon.stub() },
 	LivechatInquiry: { markInquiryActiveForPeriod: Sinon.stub() },
 	LivechatRooms: {
 		getVisitorActiveForPeriodUpdateQuery: Sinon.stub(),
@@ -21,8 +21,8 @@ const { markRoomResponded } = proxyquire.load('../../../../../app/livechat/serve
 
 describe('markRoomResponded', () => {
 	beforeEach(() => {
-		models.LivechatVisitors.isVisitorActiveOnPeriod.reset();
-		models.LivechatVisitors.markVisitorActiveForPeriod.reset();
+		models.LivechatContacts.isContactActiveOnPeriod.reset();
+		models.LivechatContacts.markContactActiveForPeriod.reset();
 		models.LivechatInquiry.markInquiryActiveForPeriod.reset();
 		models.LivechatRooms.getVisitorActiveForPeriodUpdateQuery.reset();
 		models.LivechatRooms.getAgentLastMessageTsUpdateQuery.reset();
@@ -72,7 +72,7 @@ describe('markRoomResponded', () => {
 
 		await markRoomResponded(message, room, {});
 
-		expect(models.LivechatVisitors.markVisitorActiveForPeriod.calledOnce).to.be.true;
+		expect(models.LivechatContacts.markContactActiveForPeriod.calledOnce).to.be.true;
 	});
 
 	it('should try to mark inquiry as active for current period when room.v.activity doesnt include current period', async () => {