+++ packages/models/src/models/LivechatContacts.ts
@@ -79,6 +79,10 @@ export class LivechatContactsRaw extends BaseRaw<ILivechatContact> implements IL
 				unique: false,
 			},
 			{
+				key: { activity: 1 },
+				sparse: true,
+			},
+			{
 				key: { channels: 1 },
 				unique: false,
 			},
@@ -279,6 +283,34 @@ export class LivechatContactsRaw extends BaseRaw<ILivechatContact> implements IL
 		const updatedContact = await this.findOneAndUpdate({ _id: contactId }, { $addToSet: { emails: { address: email } } });
 
 		return updatedContact;
+	}
+
+	isContactActiveOnPeriod(visitor: ILivechatContactVisitorAssociation, period: string): Promise<number> {
+		const query = {
+			...this.makeQueryForVisitor(visitor),
+			activity: period,
+		};
+
+		return this.countDocuments(query);
+	}
+
+	markContactActiveForPeriod(visitor: ILivechatContactVisitorAssociation, period: string): Promise<UpdateResult> {
+		const update = {
+			$push: {
+				activity: {
+					$each: [period],
+					$slice: -12,
+				},
+			},
+		};
+
+		return this.updateOne(this.makeQueryForVisitor(visitor), update);
+	}
+
+	countContactsOnPeriod(period: string): Promise<number> {
+		return this.countDocuments({
+			activity: period,
+		});
 	}
 
 	countByContactInfo({ contactId, email, phone }: { contactId?: string; email?: string; phone?: string }): Promise<number> {